server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
openai==0.27.8
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
python-docx==0.8.11
reportlab==4.0.4
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
alembic==1.12.0
redis==4.6.0
pytest==7.4.0
Pillow==10.0.0
PyMuPDF==1.23.0
openpyxl==3.1.2
Werkzeug==2.3.7
flask-socketio==5.3.6
python-socketio==5.8.0
matplotlib==3.7.2
seaborn==0.12.2
python-pptx==0.6.21
